Changed FloorsService to FloorService, its path from floors to floor, getAll to list, getOne to retrieve; Updated imports to new path and class; Added docstrings to Floor related services and model; Update BlockNo docstring
