+++ web/src/app/components/parkings/add-parking/add-parking.component.ts
@@ -6,9 +6,9 @@ import { LoadingBarService } from '@ngx-loading-bar/core';
 import { NotifyService } from 'src/app/shared/handlers/notify/notify.service';
 
 import { Block } from 'src/app/shared/services/block/block.model';
-import { Floor } from 'src/app/shared/services/floors/floors.model';
+import { Floor } from 'src/app/shared/services/floor/floor.model';
 import { BlockService } from 'src/app/shared/services/block/block.service';
-import { FloorsService } from 'src/app/shared/services/floors/floors.service';
+import { FloorService } from 'src/app/shared/services/floor/floor.service';
 import { ParkingService } from 'src/app/shared/services/parking/parking.service';
 
 @Component({
@@ -52,7 +52,7 @@ export class AddParkingComponent implements OnInit, OnDestroy {
     private loadingBar: LoadingBarService,
     private notifySvc: NotifyService,
     private blockSvc: BlockService,
-    private floorSvc: FloorsService,
+    private floorSvc: FloorService,
     private parkingSvc: ParkingService
   ) { }
 
@@ -82,7 +82,7 @@ export class AddParkingComponent implements OnInit, OnDestroy {
 
     this.subscription.add(forkJoin([
       this.blockSvc.list(),
-      this.floorSvc.getAll()
+      this.floorSvc.list()
     ]).subscribe({
       next: () => {
         this.loadingBar.useRef('http').complete()
